+++ server/web/middlewares/authenticate-bearer.js
@@ -3,7 +3,14 @@
 var env = require('gitter-web-env');
 var logger = env.logger;
 var StatusError = require('statuserror');
+var escapeStringRegexp = require('escape-string-regexp');
+var unauthorizedRedirectMap = require('../../utils/unauthorized-redirect-map');
 var oauthService = require('../../services/oauth-service');
+
+var unauthorizedUrlRedirectRegexList = Object.keys(unauthorizedRedirectMap).map((redirectKey) => {
+  const redirectUrl = unauthorizedRedirectMap[redirectKey];
+  return new RegExp(`${escapeStringRegexp(redirectUrl)}(\\/(\\?.*)?)?$`);
+});
 
 function getAccessToken(req) {
 
@@ -44,13 +51,20 @@ function getAccessToken(req) {
   }
 
 }
+
 /**
  *
  */
 module.exports = function(req, res, next) {
   /* No access token? Continue! */
   var accessToken = getAccessToken(req);
-  if(!accessToken) return next();
+
+  // Avoid a redirect loop even when someone is forcing a token via
+  // `?access_token=xxxtoken` query parameter or `Authorization: bearer xxxtoken` header
+  var alreadyOnUnauthorizedUrl = unauthorizedUrlRedirectRegexList.some((unauthorizedRe) => {
+    return req.url.match(unauthorizedRe);
+  });
+  if(!accessToken || alreadyOnUnauthorizedUrl) return next();
 
   return oauthService.validateAccessTokenAndClient(accessToken)
     .then(function(tokenInfo) {